+++ appenders/src/test/java/com/kdgregory/log4j/aws/TestKinesisAppender.java
@@ -19,7 +19,9 @@ import static net.sf.kdgcommons.test.StringAsserts.*;
 import java.net.URL;
 import java.util.ArrayList;
 import java.util.Arrays;
+import java.util.HashSet;
 import java.util.List;
+import java.util.Set;
 
 import org.junit.After;
 import org.junit.Before;
@@ -33,6 +35,7 @@ import org.apache.log4j.PropertyConfigurator;
 import org.apache.log4j.helpers.LogLog;
 
 import net.sf.kdgcommons.lang.StringUtil;
+import net.sf.kdgcommons.test.StringAsserts;
 
 import com.amazonaws.services.kinesis.AmazonKinesis;
 import com.amazonaws.services.kinesis.model.*;
@@ -709,4 +712,43 @@ public class TestKinesisAppender
                                                                     BinaryUtils.copyAllBytesFrom(staticFactoryMock.putRecordsSourceRecords.get(0).getData()),
                                                                     "UTF-8"));
     }
+
+
+    @Test
+    public void testRandomPartitionKey() throws Exception
+    {
+        initialize("TestKinesisAppender/testRandomPartitionKey.properties");
+
+        final Set<String> partitionKeys = new HashSet<String>();
+        MockKinesisClient mockClient = new MockKinesisClient()
+        {
+            @Override
+            protected PutRecordsResult putRecords(PutRecordsRequest request)
+            {
+                for (PutRecordsRequestEntry entry : request.getRecords())
+                {
+                    partitionKeys.add(entry.getPartitionKey());
+                }
+                return super.putRecords(request);
+            }
+        };
+
+        appender.setThreadFactory(new DefaultThreadFactory());
+        appender.setWriterFactory(mockClient.newWriterFactory());
+
+        for (int ii = 0 ; ii < 10 ; ii++)
+        {
+            logger.debug("example message");
+        }
+        mockClient.allowWriterThread();
+
+        // it's possibly but unlikely for this to fail -- we could randomly get same value
+        assertEquals("number of partition keys", 10, partitionKeys.size());
+
+        for (String key : partitionKeys)
+        {
+            StringAsserts.assertRegex("partition key is 8 digits (was: " + key + ")", "\\d{8}", key);
+        }
+    }
+
 }